implemented movement on heading

+++ inc/interfaces/Coord.hpp
@@ -3,10 +3,11 @@
 #include <cstdint>
 
 struct Coord final {
-  int64_t x{0};
-  int64_t y{0};
-  int64_t z{0};
+  long double x{0};
+  long double y{0};
+  long double z{0};
 
-  Coord(int64_t x, int64_t y, int64_t z = 0) noexcept : x(x), y(y), z(z) {}
+  Coord(long double x, long double y, long double z = 0) noexcept
+      : x(x), y(y), z(z) {}
   Coord() noexcept = default;
 };
